+++ tt_metal/hw/inc/api/compute/experimental/matmul_custom.h
@@ -10,7 +10,7 @@
 #endif
 #ifdef TRISC_UNPACK
 #include "llk_unpack_AB_matmul_api.h"
-#include "llk_unpack_AB_matmul_custom_api.h"
+#include "experimental/llk_unpack_AB_matmul_custom_api.h"
 #endif
 #ifdef TRISC_PACK
 #include "llk_pack_api.h"
