+++ tt_metal/hw/inc/api/compute/reduce_custom.h
@@ -127,7 +127,7 @@ ALWI void reduce_block_max_row(uint32_t icb, uint32_t icb_scaler, uint32_t row_s
 template <uint32_t block_ct_dim, bool respect_trigger = false>
 ALWI void reduce_block_max_row_reinit_short() {
     UNPACK((llk_unpack_AB_reduce_block_max_row_init<block_ct_dim, DST_ACCUM_MODE, respect_trigger>()));
-    MATH((llk_math_reduce_block_max_row_reinit_with_mop<block_ct_dim, DST_ACCUM_MODE>()));
+    MATH((llk_math_reduce_block_max_row_reinit_with_mop<block_ct_dim>()));
     PACK((llk_pack_reduce_mask_config<false, ReduceDim::REDUCE_ROW>()));
 }
